Seccion de administracion y algunos ajustes

+++ app/Http/Controllers/ArchivoController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use File;
 use Redirect;
 use App\Models\Archivo;
+use App\Models\User;
 use Illuminate\Support\Facades\DB;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -17,7 +18,14 @@ class ArchivoController extends Controller
     }
     public function index(Request $request){
         $request->user()->authorizeRoles(['user', 'admin']);
-        return view('subir_archivo');
+        $user_id=0;
+        $user_name='';
+        if(Auth::user()->hasRole('admin') and $request->user_id){
+            $user=User::find($request->user_id);
+            $user_id=$user->id;
+            $user_name=$user->name;
+        }
+        return view('subir_archivo',compact('user_id','user_name'));
     }
 
     public function subir_archivo(Request $request){
@@ -25,9 +33,24 @@ class ArchivoController extends Controller
     }
 
     public function historial_archivo(Request $request){
+        $user=Auth::user();
+        $usuarios=false;
         $request->user()->authorizeRoles(['user', 'admin']);
-        $archivos=Archivo::paginate(10);
-        return view('historial_archivo',compact('archivos'));
+        $archivos=Archivo::where('user_id',$user->id)->paginate(10);
+        return view('historial_archivo',compact('archivos','usuarios'));
+    }
+
+    public function historial_archivo_admin(Request $request){
+        $request->user()->authorizeRoles(['admin']);
+        $archivos=DB::table('archivos');
+        if($request->user_id){
+            $archivos=$archivos->where('user_id',$request->user_id);
+            $archivos=$archivos->paginate(10)->appends('user_id',request('user_id'));
+        }else{
+            $archivos=$archivos->paginate(10);
+        }
+        $usuarios=User::all();
+        return view('historial_archivo',compact('archivos','usuarios'));
     }
 
     /**
@@ -47,7 +70,11 @@ class ArchivoController extends Controller
                 $archivo->nombre=$request->file->getClientOriginalName();
                 $archivo->nombre_archivo=$filename;
                 $archivo->peso=$peso;
-                $archivo->user_id=$user->id;
+                if(Auth::user()->hasRole('admin') and $request->user_id){
+                    $archivo->user_id=$request->user_id;
+                }else{
+                    $archivo->user_id=$user->id;
+                }
                 $archivo->save();
                 return redirect::to('/archivo')->with('success','El archivo '.$request->file->getClientOriginalName().' fue subido de forma correcta.');
             }catch(Exception $e){
@@ -66,9 +93,17 @@ class ArchivoController extends Controller
 
         $user=Auth::user();
         $archivo=Archivo::find($request->id);
-        $path=public_path('/files/'.$user->name.'_'.$user->id.'/'.$archivo->nombre_archivo);
-        File::delete($path);
-        $archivo->delete();
-        return redirect::to('/historial')->with('success','El archivo '.$archivo->nombre.' fue eliminado de forma correcta.');
+        if($archivo->user_id==$user->id or Auth::user()->hasRole('admin')){
+            $path=public_path('/files/'.$user->name.'_'.$user->id.'/'.$archivo->nombre_archivo);
+            File::delete($path);
+            $archivo->delete();
+            if($request->admin){
+                return redirect::to('/historial_archivo_admin')->with('success','El archivo '.$archivo->nombre.' fue eliminado de forma correcta.');
+            }else{
+                return redirect::to('/historial')->with('success','El archivo '.$archivo->nombre.' fue eliminado de forma correcta.');
+            }
+        }
+        return redirect::to('/historial')->with('fail','Usted no posee autorización para realizar esta acción');
+        
     }
 }
